+++ lib/rubocop/formatter/clang_style_formatter.rb
@@ -10,7 +10,7 @@ module Rubocop
         offences.each do |o|
           output.printf("%s:%d:%d: %s: %s\n",
                         smart_path(file).color(:cyan), o.line, o.real_column,
-                        colored_severity_code(o), o.message)
+                        colored_severity_code(o), message(o))
 
           source_line = o.location.source_line