+++ spec/rubocop/formatter/clang_style_formatter_spec.rb
@@ -49,6 +49,39 @@ module Rubocop
                                        '^^^^',
                                        ''].join("\n")
         end
+
+        let(:file) { '/path/to/file' }
+
+        let(:offence) do
+          Cop::Offence.new(:convention, location,
+                           'This is a message.', 'CopName', corrected)
+        end
+
+        let(:location) do
+          source_buffer = Parser::Source::Buffer.new('test', 1)
+          source_buffer.source = "a\n"
+          Parser::Source::Range.new(source_buffer, 0, 1)
+        end
+
+        context 'when the offence is not corrected' do
+          let(:corrected) { false }
+
+          it 'prints message as-is' do
+            formatter.report_file(file, [offence])
+            expect(output.string)
+              .to include(': This is a message.')
+          end
+        end
+
+        context 'when the offence is automatically corrected' do
+          let(:corrected) { true }
+
+          it 'prints [Corrected] along with message' do
+            formatter.report_file(file, [offence])
+            expect(output.string)
+              .to include(': [Corrected] This is a message.')
+          end
+        end
       end
     end
   end